ColorMap handling moved to separate script (keep ModelWithPlate as small as possible, as it's too large already)
Make ButtonsClickReceiver more flexible, calls multiple IClickHandlers.
