+++ unity/Holo/Assets/GFX/Materials/DefaultModelMaterial.mat
@@ -9,7 +9,7 @@ Material:
   m_PrefabAsset: {fileID: 0}
   m_Name: DefaultModelMaterial
   m_Shader: {fileID: 4800000, guid: 96961c9614531b1488fcfc3f24d1c5da, type: 3}
-  m_ShaderKeywords: CLIPPING_ON _ALPHAPREMULTIPLY_ON
+  m_ShaderKeywords: _ALPHAPREMULTIPLY_ON
   m_LightmapFlags: 4
   m_EnableInstancingVariants: 0
   m_DoubleSidedGI: 0
