+++ unity/Holo/Assets/GFX/Materials/dataFlowMat.mat
@@ -9,7 +9,7 @@ Material:
   m_PrefabAsset: {fileID: 0}
   m_Name: dataFlowMat
   m_Shader: {fileID: 4800000, guid: 44e85b36c6b520441b4042b277e13fe9, type: 3}
-  m_ShaderKeywords: CLIPPING_ON
+  m_ShaderKeywords: 
   m_LightmapFlags: 4
   m_EnableInstancingVariants: 0
   m_DoubleSidedGI: 0
@@ -24,7 +24,7 @@ Material:
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     - _ColorMap:
-        m_Texture: {fileID: 2800000, guid: 66583419871721443ad625ce8340eeec, type: 3}
+        m_Texture: {fileID: 2800000, guid: 55020d9191c25ff4bb25fa93226153d4, type: 3}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     - _DetailAlbedoMap:
