+++ unity/Holo/Assets/Scripts/model_with_plate/ColorMap.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 53a50f7ae4a28be4b86be8fd7f79962d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
